# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fruit Store API for Russian customers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseErrorsItem = {
  field?: string;
  message?: string;
};

export interface ErrorResponse {
  message: string;
  errors?: ErrorResponseErrorsItem[];
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  CUSTOMER: "CUSTOMER",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  avatar?: string | null;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type FruitCategory = (typeof FruitCategory)[keyof typeof FruitCategory];

export const FruitCategory = {
  МЕСТНЫЕ: "МЕСТНЫЕ",
  ТРОПИЧЕСКИЕ: "ТРОПИЧЕСКИЕ",
  ОРГАНИЧЕСКИЕ: "ОРГАНИЧЕСКИЕ",
  ИМПОРТНЫЕ: "ИМПОРТНЫЕ",
} as const;

export interface Fruit {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  price: number;
  stock: number;
  category: FruitCategory;
  organic: boolean;
  imageUrl?: string | null;
  createdAt: string;
}

export interface FruitsListResponse {
  fruits: Fruit[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateFruitInputCategory =
  (typeof CreateFruitInputCategory)[keyof typeof CreateFruitInputCategory];

export const CreateFruitInputCategory = {
  МЕСТНЫЕ: "МЕСТНЫЕ",
  ТРОПИЧЕСКИЕ: "ТРОПИЧЕСКИЕ",
  ОРГАНИЧЕСКИЕ: "ОРГАНИЧЕСКИЕ",
  ИМПОРТНЫЕ: "ИМПОРТНЫЕ",
} as const;

export interface CreateFruitInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  stock: number;
  category: CreateFruitInputCategory;
  organic?: boolean;
  imageUrl?: string;
}

export interface OrderItem {
  id: number;
  fruitId: number;
  fruitName: string;
  quantity: number;
  weight: string;
  unitPrice: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  ОЖИДАНИЕ: "ОЖИДАНИЕ",
  ПОДТВЕРЖДЁН: "ПОДТВЕРЖДЁН",
  ОТПРАВЛЕН: "ОТПРАВЛЕН",
  ДОСТАВЛЕН: "ДОСТАВЛЕН",
  ОТМЕНЁН: "ОТМЕНЁН",
} as const;

export interface Order {
  id: number;
  userId: number;
  status: OrderStatus;
  totalPrice: number;
  address: string;
  items: OrderItem[];
  createdAt: string;
}

export type CreateOrderInputItemsItem = {
  fruitId: number;
  quantity: number;
  weight: string;
};

export interface CreateOrderInput {
  address: string;
  items: CreateOrderInputItemsItem[];
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  ОЖИДАНИЕ: "ОЖИДАНИЕ",
  ПОДТВЕРЖДЁН: "ПОДТВЕРЖДЁН",
  ОТПРАВЛЕН: "ОТПРАВЛЕН",
  ДОСТАВЛЕН: "ДОСТАВЛЕН",
  ОТМЕНЁН: "ОТМЕНЁН",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export type GetFruitsParams = {
  category?: GetFruitsCategory;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  organic?: boolean;
  page?: number;
  limit?: number;
};

export type GetFruitsCategory =
  (typeof GetFruitsCategory)[keyof typeof GetFruitsCategory];

export const GetFruitsCategory = {
  МЕСТНЫЕ: "МЕСТНЫЕ",
  ТРОПИЧЕСКИЕ: "ТРОПИЧЕСКИЕ",
  ОРГАНИЧЕСКИЕ: "ОРГАНИЧЕСКИЕ",
  ИМПОРТНЫЕ: "ИМПОРТНЫЕ",
} as const;