# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fruit Store API for Russian customers
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: fruits
    description: Fruit catalog operations
  - name: orders
    description: Order management operations
  - name: cart
    description: Shopping cart operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fruits:
    get:
      operationId: getFruits
      tags: [fruits]
      summary: Get all fruits with optional filtering
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [МЕСТНЫЕ, ТРОПИЧЕСКИЕ, ОРГАНИЧЕСКИЕ, ИМПОРТНЫЕ]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: organic
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of fruits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitsListResponse"

    post:
      operationId: createFruit
      tags: [fruits]
      summary: Create a new fruit (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFruitInput"
      responses:
        "201":
          description: Fruit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fruit"

  /fruits/{slug}:
    get:
      operationId: getFruitBySlug
      tags: [fruits]
      summary: Get a specific fruit by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fruit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fruit"
        "404":
          description: Fruit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fruits/{id}:
    put:
      operationId: updateFruit
      tags: [fruits]
      summary: Update a fruit (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFruitInput"
      responses:
        "200":
          description: Fruit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fruit"

    delete:
      operationId: deleteFruit
      tags: [fruits]
      summary: Delete a fruit (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fruit deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/my:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get current user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of user orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrderById
      tags: [orders]
      summary: Get order details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - message

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    Fruit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        stock:
          type: integer
        category:
          type: string
          enum: [МЕСТНЫЕ, ТРОПИЧЕСКИЕ, ОРГАНИЧЕСКИЕ, ИМПОРТНЫЕ]
        organic:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - price
        - stock
        - category
        - organic
        - createdAt

    FruitsListResponse:
      type: object
      properties:
        fruits:
          type: array
          items:
            $ref: "#/components/schemas/Fruit"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - fruits
        - total
        - page
        - totalPages

    CreateFruitInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        category:
          type: string
          enum: [МЕСТНЫЕ, ТРОПИЧЕСКИЕ, ОРГАНИЧЕСКИЕ, ИМПОРТНЫЕ]
        organic:
          type: boolean
        imageUrl:
          type: string
      required:
        - name
        - slug
        - price
        - stock
        - category

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        fruitId:
          type: integer
        fruitName:
          type: string
        quantity:
          type: integer
        weight:
          type: string
        unitPrice:
          type: number
      required:
        - id
        - fruitId
        - fruitName
        - quantity
        - weight
        - unitPrice

    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [ОЖИДАНИЕ, ПОДТВЕРЖДЁН, ОТПРАВЛЕН, ДОСТАВЛЕН, ОТМЕНЁН]
        totalPrice:
          type: number
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - status
        - totalPrice
        - address
        - items
        - createdAt

    CreateOrderInput:
      type: object
      properties:
        address:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              fruitId:
                type: integer
              quantity:
                type: integer
              weight:
                type: string
            required:
              - fruitId
              - quantity
              - weight
      required:
        - address
        - items

    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [ОЖИДАНИЕ, ПОДТВЕРЖДЁН, ОТПРАВЛЕН, ДОСТАВЛЕН, ОТМЕНЁН]
      required:
        - status