# lib/api-zod/src/generated/types/fruit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fruit Store API for Russian customers
 * OpenAPI spec version: 0.1.0
 */
import type { FruitCategory } from "./fruitCategory";

export interface Fruit {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  price: number;
  stock: number;
  category: FruitCategory;
  organic: boolean;
  imageUrl?: string | null;
  createdAt: Date;
}