# lib/api-zod/src/generated/types/fruitsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fruit Store API for Russian customers
 * OpenAPI spec version: 0.1.0
 */
import type { Fruit } from "./fruit";

export interface FruitsListResponse {
  fruits: Fruit[];
  total: number;
  page: number;
  totalPages: number;
}