# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fruit Store API for Russian customers
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  fruitId: number;
  fruitName: string;
  quantity: number;
  weight: string;
  unitPrice: number;
}